feat: Implement durable AgentRunner and background queue
Why:
The agent orchestration was previously running synchronously on the main thread, locking the UI during long LLM calls and context collection.
We needed a decoupled background queue to process tasks asynchronously while streaming progress back to the UI.

Changed:
- Created ccad_core/agent_runner.hpp/cpp implementing a standard std::thread loop with std::condition_variable queue processing.
- Modified AgentOrchestrator to delegate execution tasks to the AgentRunner instance.
- Wired AgentRunner::set_progress_callback through the orchestrator up to the Qt AgentPanel.
- Used QMetaObject::invokeMethod inside AgentPanel to safely jump thread boundaries and update UI status chips.
- Patched compiler CI/CD warnings related to type casting, M_PI, and unused lambda variables.
- Isolated test_agent_runner.cpp into its own CMake executable target.

Behavior:
Agents can now enqueue goals into the orchestrator. These run transparently in the background, updating the Qt UI iteratively as tasks finish.

Verification:
- Fixed all MSVC /W4 and -Werror warnings across C++ kernel headers.
- Passed full cmake --build pipeline including all test binaries.
- Ran ccad_agent_runner_tests locally via CTest.

Demo:
- Visual validation script (run_sprint_demo.ps1) executed successfully, proving the GUI thread remains decoupled from the orchestrator logic without rendering deadlocks.

+++ src/ccad_core/autoplace_fields.cpp
@@ -94,8 +94,8 @@ void autoplaceSymbolFields(Schematic& schematic, SchSymbol& symbol, const Autopl
     // Align to grid (50 mils = 1270000 nm)
     if (options.align_to_grid) {
       const long long grid = 1270000;
-      field.position.x.nanometers = std::round((double)field.position.x.nanometers / grid) * grid;
-      field.position.y.nanometers = std::round((double)field.position.y.nanometers / grid) * grid;
+      field.position.x.nanometers = static_cast<long long>(std::round((double)field.position.x.nanometers / grid)) * grid;
+      field.position.y.nanometers = static_cast<long long>(std::round((double)field.position.y.nanometers / grid)) * grid;
     }
     
     idx++;
